quille-backend/requirements.txt
aiohttp==3.9.4
aiohttp-cors==0.7.0
aiohttp-sse==2.2.0
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.25.2
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
multidict==6.0.5
packaging==24.0
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
